use pp_done handler instead of buffer_release handler
the buffer_release handler is not good to detect when pp is done.
it can be also called when the pp operation is canceled. At this time,
tdm_pp doesn't support the cancelation. But it will be supported soon.
Furthermore it's obvious we can use a buffer whenever the release handler
is called.

Change-Id: Ib5add0becee31663fc24afddb853dfda3a62a47a